+++ src/cmd/main.go
@@ -9,7 +9,7 @@ import (
 	"github.com/testing-cli/apitest/internal/cli"
 )
 
-const version = "0.3.4"
+const version = "0.3.5"
 
 func main() {
 	if len(os.Args) < 2 {